+++ php_wcli.h
@@ -1,6 +1,7 @@
 /* wcli extension for PHP */
 
 #include "TlHelp32.h"
+#include <libloaderapi.h>
 
 #ifndef PHP_WCLI_H
 # define PHP_WCLI_H
@@ -9,6 +10,14 @@ extern zend_module_entry wcli_module_entry;
 # define phpext_wcli_ptr &wcli_module_entry
 
 # define PHP_WCLI_VERSION "1.0.0"
+
+typedef int (__cdecl *WCLI_READ)(
+	_In_ HANDLE hConsoleInput,
+    _Out_writes_(nLength) PINPUT_RECORD lpBuffer,
+    _In_ DWORD nLength,
+    _Out_ LPDWORD lpNumberOfEventsRead,
+    _In_ USHORT wFlags
+); 
 
 ZEND_BEGIN_MODULE_GLOBALS(wcli)
     BOOL console;
@@ -21,6 +30,7 @@ ZEND_BEGIN_MODULE_GLOBALS(wcli)
     CONSOLE_SCREEN_BUFFER_INFO screen;
     CONSOLE_CURSOR_INFO cursor;
     CONSOLE_FONT_INFO font;
+    WCLI_READ ReadConsoleInputExA;
 ZEND_END_MODULE_GLOBALS(wcli)
 
 ZEND_EXTERN_MODULE_GLOBALS(wcli)
@@ -30,5 +40,14 @@ ZEND_EXTERN_MODULE_GLOBALS(wcli)
 # if defined(ZTS) && defined(COMPILE_DL_WCLI)
 ZEND_TSRMLS_CACHE_EXTERN()
 # endif
+
+
+#ifndef CONSOLE_READ_NOREMOVE
+#define CONSOLE_READ_NOREMOVE   0x0001
+#endif
+
+#ifndef CONSOLE_READ_NOWAIT
+#define CONSOLE_READ_NOWAIT     0x0002
+#endif
 
 #endif	/* PHP_WCLI_H */